Updated project to use separate workspaces for base and app
there are now two crates to represent the base sw and app sw.

base sw interacts with hardware and compiles to cortex-m targets.

app sw is pure with no hw dependency and can be built to any
target. This means that tests that are written for app sw can run on
the default rust target of the development machine to speed them up.

Also add rust implementation of goertzel algorithm and started work on tests for it in app crate